+++ CHANGES.md
@@ -4,6 +4,7 @@
 - cleaned ReplacementLoadFromPath() up
 ## Furniture
 - added Furniture Handler -> Callback when Furniture gets loaded, which allows us to add custum scripts to the prefabs
+- fixed Furniture stacking issues
 
 # v0.1.0-alpha (July 4th 2024) - Preview
 ## General